// src/API.ts
export type CreateTodoInput = {
  id?: string | null,
  userId: string,
  name: string,
  description?: string | null,
};
// ...
export type Todo = {
  __typename: "Todo",
  id: string,
  userId: string,
  name: string,
  description?: string | null,
  createdAt: string,
  updatedAt: string,
};